+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ### Improvements
 - Renamed 'velocity' (= a vector) to 'speed' (= a value) where appropriate
 - When starting a new recording the user aircraft is now updated in the logbook as soon as it becomes available
+- The aircraft table in the Formation module is now updated by row (instead of the entire table) upon time offset changes
 
 ### Bug Fixes
 - The window is properly resized when in minimal mode and the default non-essential button visibility is changed in the settings
@@ -22,6 +23,7 @@
 
 ### Under The Hood
 - The modules (Logbook, Formation, Location) are now implemented as plugins
+- The modules are [topologically sorted](https://en.wikipedia.org/wiki/Topological_sorting), based on their defined dependencies
 - Sky Dolly now also compiles and links with the Microsoft Visual Studio 2022 C++ compiler (MSVC)
 - The code is now also [analysed](https://github.com/till213/SkyDolly/actions) for programming / security flaws with the [Microsoft C++ Code Analysis Action](https://github.com/marketplace/actions/microsoft-c-code-analysis-action)